Actualización en modulo de control escolar  Seguimiento caso crítico especialista: Se habilitó únicamente al área  de psicopedagogia la oportunidad de cerrar un caso crítico  Registro de Plantilla de Plan de Acción Tutoría: Se habilitó al área de psicopedagogia una herramienta para realizar el registro de plantillas de Plan de Acción Tutorial.  Plan de Acción Tutorial: Se agregó la función de tomar la plantilla del área de psicopedagogia de manera automática, para así evitar el registro completo por parte del tutor  Tutoría Individual: Se agregó un candado y validación en la liberación de casos críticos cuando se encuentren en seguimiento por el especialista psicopedagogia  Se eliminó en vistas el uso de SortBy debido a que en la actualización de datos con ajax se tomaban valores errores  Validación en el acceso del estudiante, se quitarón las conversiones del init pasando a modelo

+++ src/main/java/mx/edu/utxj/pye/sgi/util/ServicioArchivos.java
@@ -105,8 +105,10 @@ public class ServicioArchivos implements Serializable{
     
     public static void eliminarArchivo(String ruta){
         try {
-            Path path = Paths.get(ruta);
-            Files.deleteIfExists(path);
+            if (!ruta.equals("") || ruta != null) {
+                Path path = Paths.get(ruta);
+                Files.deleteIfExists(path);
+            }
         } catch (IOException ex) {
             Logger.getLogger(ServicioArchivos.class.getName()).log(Level.SEVERE, null, ex);
         }
